requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==24.2
PyJWT==2.10.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
